+++ src/main/java/kitchenpos/products/tobe/domain/ProductRepository.java
@@ -1,6 +1,5 @@
-package kitchenpos.products.infrastructure.dao;
+package kitchenpos.products.tobe.domain;
 
-import kitchenpos.products.tobe.domain.Product;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface ProductRepository extends JpaRepository<Product, Long> {
